Add http status  test to content app
