+++ content/webapp/jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  transformIgnorePatterns: ['node_modules(?!/@weco(?!.*node_modules))']
+};
